Move back to mr-clojure group
Turns out you can't nest like that, plus even if you could you really
would be better off with lein new mr-clojure rather than having to put
in the additional group. Whilst it would be nice to be in the mixradio
clojars group I think this is the lesser of the two evils. lein new
mixradio seems a little off. It also precludes us from creating more
templates in the future.