Web server can now be started and stopped without REPL restarts...
Remove metrics in the stop function so that we can restart the web server from within the REPL, making our services fully reloadable.

I've created a new function `kill-service` which is called when the service shuts down. All I've done here is pull `shutdown-agents` out of the `stop` fn as it completely destroys any running REPL; in most cases this is undesirable. We only want out agent threads to be killed off if we are restarting the service in production.

Radix 1.0.12